+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  
 ## [Unreleased]
+### Changed
+- **BREAKING**: Renamed property `debugMode` to `debug` in `Agent` constructor. \
+`debug` will also accept a callback function that is called for all log messages.
+
+---
 
 ## [0.11.19] - 2021-04-19
 ### Added